+++ src/index.js
@@ -1,3 +1,4 @@
 export { default } from "./layout.js";
 export { default as CloudLayout } from "./layout.js";
+export { archimedeanStrategy, rectangularStrategy, noneStrategy } from "./layout.js";
 export { default as CloudSprite } from "./sprite.js";